delete duplicate services and mappers delete duplicate filters
